requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
djangorestframework
drf-spectacular
